Changed User.id int to Long.

+++ src/main/java/com/softserve/betterlearningroom/controller/AuthController.java
@@ -51,7 +51,7 @@ public class AuthController {
     }
 
     @PutMapping("/users/{id}")
-    public ResponseEntity<UserDTO> updateUser(@RequestBody @Valid SaveUserRequest request, @PathVariable int id) throws UserAlreadyExistsException {
+    public ResponseEntity<UserDTO> updateUser(@RequestBody @Valid SaveUserRequest request, @PathVariable Long id) throws UserAlreadyExistsException {
         UserDTO updatedUser = authService.updateUser(request, id);
         return ResponseEntity.ok().body(updatedUser);
     }
